Setup classes for native paintworklet
This CL creates classes to build the infra of native paint worklet.
Specifically, we create a NativePaintWorklet class, and when
we paint the background color, we ask the NativePaintWorklet
to do the paint. Because NativePaintWorklet lives under the
modules/ and paint part is in core/, we need some other classes
to be there so that the paint can call NativePaintWorklet APIs.

Design doc here:
https://docs.google.com/document/d/1usCnwWs8HsH5FU_185q6MsrZehFmpl5QgbbB4pvHIjI/edit

Bug: 1138997
Change-Id: Ifdabf3ce5226ee4cacac03eec46029e74c3fa9c6